requirements.txt
requests==2.31.0
scapy==2.5.0
python-nmap==0.7.1